extract ebpf load config
